Apply all updates for lib_ke_protocol

+++ Core/Src/main.c
@@ -517,7 +517,7 @@ static void esp32_reset( HOST_PWR_STATE state )
 		HAL_GPIO_WritePin(ESP32_RESET_N_GPIO_Port, ESP32_RESET_N_Pin, GPIO_PIN_RESET);
 }
 
-static int esp32_tx( uint8_t *data, uint8_t len )
+static int esp32_tx( uint8_t *data, uint32_t len )
 {
     if( HAL_UART_Transmit_IT(ESP32_UART, data, len) == HAL_OK )
         return 1;
